Add a smoke tests and local development details (#3)

+++ bundler_utils.rb
@@ -1,21 +1,23 @@
+# frozen_string_literal: true
+
 require 'fileutils'
 require 'rest-client'
 require 'yaml'
 
 def with_gem_auth(is_private, &block)
   if is_private
-    puts "Using gems.salsify.com authentication."
+    puts 'Using gems.salsify.com authentication.'
     with_gems_salsify_auth(&block)
   else
-    puts "Using rubygems.org authentication."
+    puts 'Using rubygems.org authentication.'
     with_ruby_gems_auth(&block)
   end
 end
 
 def with_ruby_gems_auth
   with_credentials_file do |creds_file|
-    api_key = ENV['RUBYGEMS_API_KEY'] or exit_with_error("RUBYGEMS_API_KEY not set!")
-    File.write(creds_file, { rubygems_api_key: api_key }.to_yaml)
+    api_key = ENV['RUBYGEMS_API_KEY'] or exit_with_error('RUBYGEMS_API_KEY not set!')
+    File.write(creds_file, { rubygems_api_key: api_key }.to_yaml) if update_config?
 
     yield
   end
@@ -23,15 +25,15 @@ end
 
 def with_gems_salsify_auth
   with_credentials_file do |creds_file|
-    auth = ENV['ARTIFACTORY_AUTH_STRING'] or exit_with_error("ARTIFACTORY_AUTH_STRING not set!")
+    auth = ENV['ARTIFACTORY_AUTH_STRING'] or exit_with_error('ARTIFACTORY_AUTH_STRING not set!')
     api_key = RestClient.get("https://#{auth}@gems.salsify.com/api/v1/api_key").body
-    File.write(creds_file, { gems_salsify_com: api_key }.to_yaml)
+    File.write(creds_file, { gems_salsify_com: api_key }.to_yaml) if update_config?
 
-    system("bundle config gems.salsify.com #{auth}", exception: true)
+    system("bundle config gems.salsify.com #{auth}", exception: true) if update_config?
 
     yield
   ensure
-    system('bundle config gems.salsify.com')
+    system('bundle config gems.salsify.com') if update_config?
   end
 end
 
@@ -45,7 +47,7 @@ def with_credentials_file
 
   yield creds_file
 ensure
-  FileUtils.remove_dir(gem_dir, true)
+  FileUtils.remove_dir(gem_dir, true) if update_config?
 end
 
 def release_gem
@@ -54,4 +56,8 @@ def release_gem
   as_git_user(name: 'github-actions', email: 'github-actions@user.noreply.github.com') do
     system('bundle exec rake release', exception: true)
   end
+end
+
+def update_config?
+  ENV['ACTION_ENV'] == 'production'
 end
